Issue #48 - Change comparison counts type to int

+++ src/main/java/org/scify/jedai/datamodel/AbstractBlock.java
@@ -31,7 +31,7 @@ public abstract class AbstractBlock implements Serializable {
     
     protected int blockIndex;
     
-    protected float comparisons;
+    protected int comparisons;
     protected float entropy;
     protected float utilityMeasure;
             
@@ -53,7 +53,7 @@ public abstract class AbstractBlock implements Serializable {
         return entropy;
     }
 
-    public float getNoOfComparisons() {
+    public int getNoOfComparisons() {
         return comparisons;
     }
