+++ src/test/java/org/scify/jedai/configuration/StepByStepGridConfigurationCCER.java
@@ -47,9 +47,9 @@ import org.scify.jedai.utilities.datastructures.BilateralDuplicatePropagation;
  */
 public class StepByStepGridConfigurationCCER {
 
-    static float getTotalComparisons(List<AbstractBlock> blocks) {
-        float originalComparisons = 0;
-        originalComparisons = blocks.stream().map(AbstractBlock::getNoOfComparisons).reduce(originalComparisons, Float::sum);
+    static int getTotalComparisons(List<AbstractBlock> blocks) {
+        int originalComparisons = 0;
+        originalComparisons = blocks.stream().map(AbstractBlock::getNoOfComparisons).reduce(originalComparisons, Integer::sum);
         System.out.println("Original comparisons\t:\t" + originalComparisons);
         return originalComparisons;
     }
@@ -104,7 +104,7 @@ public class StepByStepGridConfigurationCCER {
             // local optimization of Block Building 
             float bestA = 0;
             int bestIteration = 0;
-            float originalComparisons = ((float) profiles1.size()) * profiles2.size();
+            int originalComparisons = profiles1.size() * profiles2.size();
             for (int j = 0; j < bb.getNumberOfGridConfigurations(); j++) {
                 bb.setNumberedGridConfiguration(j);
                 final List<AbstractBlock> originalBlocks = bb.getBlocks(profiles1, profiles2);
