+++ src/test/java/org/scify/jedai/configuration/StepByStepGridConfigurationDER.java
@@ -47,9 +47,9 @@ import java.util.List;
  */
 public class StepByStepGridConfigurationDER {
 
-    static float getTotalComparisons(List<AbstractBlock> blocks) {
-        float originalComparisons = 0;
-        originalComparisons = blocks.stream().map(AbstractBlock::getNoOfComparisons).reduce(originalComparisons, Float::sum);
+    static int getTotalComparisons(List<AbstractBlock> blocks) {
+        int originalComparisons = 0;
+        originalComparisons = blocks.stream().map(AbstractBlock::getNoOfComparisons).reduce(originalComparisons, Integer::sum);
         System.out.println("Original comparisons\t:\t" + originalComparisons);
         return originalComparisons;
     }
@@ -89,7 +89,7 @@ public class StepByStepGridConfigurationDER {
             // local optimization of Block Building 
             float bestA = 0;
             int bestIteration = 0;
-            float originalComparisons = ((float) profiles.size()) * (profiles.size()-1.0f) / 2.0f;
+            int originalComparisons = profiles.size() * (profiles.size() - 1) / 2;
             for (int j = 0; j < bb.getNumberOfGridConfigurations(); j++) {
                 bb.setNumberedGridConfiguration(j);
                 final List<AbstractBlock> originalBlocks = bb.getBlocks(profiles);
